requirements.txt
google-generativeai
python-dotenv
sqlmodel
psycopg2-binary
fastapi
uvicorn[standard]
httpx